+++ tests/test_functional.py
@@ -127,3 +127,10 @@ class FunctionalTest(TestCase):
         sonar_way_key = next(p["key"] for p in sonar_way_ts["profiles"])
         sonar_way_rule = self._search_rules(qprofile=sonar_way_key, rule_key="swift:S104")
         self.assertEqual(0, sonar_way_rule["total"])  # is not listed as a disabled key
+
+    def test_autodetect_ai_generated_code_disabled(self):
+        """Check that the autodetect AI-generated code feature is disabled."""
+        settings_api = f"{self.sonar_api}/settings"
+        autodetect_params = dict(key="sonar.autodetect.ai.code")
+        ai_autodetect_setting = requests.get(settings_api, auth=self.sonar_auth, params=autodetect_params)
+        self.assertFalse(ai_autodetect_setting)